更新Schema

+++ volans-flink-sink/src/main/scala/com/haizhi/volans/sink/component/RestEsSink.scala
@@ -3,7 +3,7 @@ package com.haizhi.volans.sink.component
 import java.util
 
 import com.haizhi.volans.common.flink.base.scala.util.JSONUtils
-import com.haizhi.volans.sink.config.constant.{CoreConstants, Keys, StoreType}
+import com.haizhi.volans.sink.config.constant.{CoreConstants, Keys, OperationMode, StoreType}
 import com.haizhi.volans.sink.config.schema.SchemaVo
 import com.haizhi.volans.sink.config.store.StoreEsConfig
 import com.haizhi.volans.sink.server.EsDao
@@ -23,6 +23,7 @@ class RestEsSink(override var storeType: StoreType,
   private val logger = LoggerFactory.getLogger(classOf[RestEsSink])
   override var uid: String = "ES"
   private var esDao: EsDao = _
+  private var operationMode: OperationMode = _
 
   override def open(parameters: Configuration): Unit = {
     esDao = new EsDao()
@@ -30,35 +31,51 @@ class RestEsSink(override var storeType: StoreType,
     esDao.initClient(storeConfig)
     // 创建index、type
     esDao.createTableIfNecessary(storeConfig)
+    this.operationMode = OperationMode.findStoreType(schemaVo.operation.mode)
   }
 
   override def invoke(elements: Iterable[String], context: SinkFunction.Context[_]): Unit = {
-    // filter elements
-    val filteredTuple: List[(String, java.util.Map[String, Object], Boolean)] =
-      elements.map(record => {
-        val recordMap = JSONUtils.jsonToJavaMap(record)
-        validateAndMerge(recordMap)
-        val filterFlag = recordMap.get(schemaVo.operation) != null && CoreConstants.OPERATION_DELETE.equalsIgnoreCase(recordMap.get(schemaVo.operation).toString)
-        recordMap.remove(schemaVo.operation)
-        // (source string,converted map,filter flag)
-        (JSONUtils.toJson(recordMap), recordMap, filterFlag)
-      }
-      ).toList
+    var deleteList: List[String] = null
+    var upsertList: List[java.util.Map[String, Object]] = null
 
+    operationMode match {
+      case OperationMode.MIX =>
+        // filter elements
+        val filteredTuple: List[(String, java.util.Map[String, Object], Boolean)] =
+          elements.map(record => {
+            val recordMap = JSONUtils.jsonToJavaMap(record)
+            validateAndMerge(recordMap)
+            val filterFlag = recordMap.get(schemaVo.operation) != null && CoreConstants.OPERATION_DELETE.equalsIgnoreCase(recordMap.get(schemaVo.operation).toString)
+            recordMap.remove(schemaVo.operation)
+            // (source string,converted map,filter flag)
+            (JSONUtils.toJson(recordMap), recordMap, filterFlag)
+          }
+          ).toList
+        deleteList = filteredTuple.filter(_._3).map(_._2.get(Keys.OBJECT_KEY).toString)
+        upsertList = filteredTuple.filter(!_._3).map(_._2)
+      case OperationMode.DELETE =>
+        deleteList = elements.map(record => {
+          val recordMap = JSONUtils.jsonToJavaMap(record)
+          validateAndMerge(recordMap)
+          recordMap.get(Keys.OBJECT_KEY).toString
+        }).toList
+      case OperationMode.UPSERT =>
+        upsertList = elements.map(record => {
+          val recordMap = JSONUtils.jsonToJavaMap(record)
+          validateAndMerge(recordMap)
+          recordMap
+        }).toList
+    }
     // Delete List
-    val deleteList = filteredTuple.filter(_._3).map(_._2.get(Keys.OBJECT_KEY).toString)
-    if(deleteList.size > 0){
+    if (deleteList != null && deleteList.size > 0) {
       logger.debug(s"es delete list: $deleteList")
-      esDao.bulkDelete(deleteList,storeConfig)
+      esDao.bulkDelete(deleteList, storeConfig)
     }
-
     // Upsert List
-    val upsertList = filteredTuple.filter(!_._3).map(_._2)
-    if(upsertList.size > 0){
+    if (upsertList != null && upsertList.size > 0) {
       logger.debug(s"es upsert list: $upsertList")
-      esDao.bulkUpsert(upsertList,storeConfig)
+      esDao.bulkUpsert(upsertList, storeConfig)
     }
-
   }
 
   /**
